requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
beautifulsoup4==4.14.2
bert-score==0.3.13
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
decorator==5.2.1
deep-translator==1.11.4
dill==0.3.8
evaluate==0.4.6
exceptiongroup==1.3.0
fastapi==0.119.0
ffmpy==0.6.3
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
gradio==3.39.0
gradio_client==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.7
importlib_resources==6.5.2
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
linkify-it-py==2.0.3
llvmlite==0.43.0
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.8.0
nltk==3.9.2
numba==0.60.0
numpy==1.25.1
openai-whisper==20250625
orjson==3.11.3
packaging==25.0
pandas==2.3.3
Pillow==9.5.0
pip==25.2
platformdirs==4.4.0
pooch==1.8.2
propcache==0.4.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
ruff==0.14.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
starlette==0.48.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.13.3
tomlkit==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.33.1
typer==0.19.2
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.37.0
websockets==11.0.3
wheel==0.45.1
win-inet-pton==1.1.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0